Fix: Update Vercel configuration and build settings
