+++ next.config.mjs
@@ -1,0 +1,39 @@
+// @ts-check
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  // Disable React Strict Mode for now to prevent duplicate API calls in development
+  reactStrictMode: false,
+  
+  // Enable production browser source maps
+  productionBrowserSourceMaps: true,
+  
+  // Configure images
+  images: {
+    // Enable image optimization
+    unoptimized: false,
+    // Allow images from these domains
+    domains: [
+      'localhost',
+      '127.0.0.1',
+      'images.unsplash.com',
+      'atmxiuindyakhqmdqzal.supabase.co',
+    ],
+  },
+  
+  // Enable experimental features
+  experimental: {
+    // Enable server actions
+    serverActions: true,
+    // Enable server components external packages
+    serverComponentsExternalPackages: ['@prisma/client', 'bcryptjs'],
+  },
+  
+  // Webpack configuration
+  webpack: (config, { isServer }) => {
+    // Important: return the modified config
+    return config;
+  },
+};
+
+export default nextConfig;
