+++ package.json
@@ -8,7 +8,11 @@
     "start": "next start",
     "lint": "next lint",
     "vercel-build": "next build",
-    "vercel-dev": "next dev"
+    "vercel-dev": "next dev",
+    "postinstall": "prisma generate"
+  },
+  "prisma": {
+    "schema": "./prisma/schema.prisma"
   },
   "dependencies": {
     "@dnd-kit/core": "^6.3.1",
